filtering detections and qa'ing data
creating basic data sumamaries to qa data entry. Either there is a lot of noise in the acoustic detections, or there are data entry and wrangling errors as fish released downstream are being detected upstream at both the site they were released at and at alternative sites which is impossible.